requirements.txt
Django==4.2.7
django-environ==0.11.2
django-tinymce==3.6.1
django-extensions==3.2.1
ipdb==0.13.13
Pillow==10.1.0
psycopg2==2.9.9
